Close to restoring functionality after adding multiple signature systems and repairing some of the bugs, more to go

+++ src/main/java/com/tokera/ate/dao/msg/MessageKeyType.java
@@ -4,13 +4,13 @@ package com.tokera.ate.dao.msg;
 
 public final class MessageKeyType {
   private MessageKeyType() { }
-  public static final int unknown = 0;
-  public static final int ntru = 1;
-  public static final int ntru_sign = 2;
-  public static final int qtesla = 3;
-  public static final int newhope = 4;
-  public static final int xmss = 5;
-  public static final int xmssmt = 6;
+  public static final short unknown = 0;
+  public static final short ntru = 1;
+  public static final short ntru_sign = 2;
+  public static final short qtesla = 3;
+  public static final short newhope = 4;
+  public static final short xmss = 5;
+  public static final short xmssmt = 6;
 
   public static final String[] names = { "unknown", "ntru", "ntru_sign", "qtesla", "newhope", "xmss", "xmssmt", };
